==> AHLE Dashboard/Data and Processing Code/CODE/1u_extract_usda_feedintakeandprice.py <==
#%% About
'''
This reads raw data from USDA feed and costs
'''
#%% Import

# os.path.join(RAWDATA_FOLDER ,'Broiler, Turkey, and Egg feed costs and market prices.xlsx')
# os.path.join(RAWDATA_FOLDER ,'usda_feedprice_broilers_chicks_hogs_2011_2014.csv')
