==> AHLE Dashboard/Data and Processing Code/CODE/1z_extract_othervars.py <==
# ----------------------------------------------------------------------------
# Program organization idea: Each variable from data organizer that doesn't
# come from a major source will get its own program file.
#
# E.g. this program covers Cost of Chicks placed, and brings in all sources
# or single-number estimates
# ----------------------------------------------------------------------------
